functions/requirements.txt
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.18.0
firebase-functions==0.1.0
requests==2.32.3
